lou/backend/pyproject.toml
[tool.poetry]
name = "lou-backend"
version = "1.0.0"
description = "Lou — Intelligent Playbook Engine for Siemens MHL 2026"
authors = ["Team Aivokat"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
sqlmodel = "^0.0.19"
litellm = "^1.40.0"
sentence-transformers = "^3.0.0"
chromadb = "^0.5.0"
python-docx = "^1.1.0"
openpyxl = "^3.1.0"
python-multipart = "^0.0.9"
aiofiles = "^23.2.0"
pypdf = "^4.2.0"
python-dotenv = "^1.0.0"
